# Optimize parsing microseconds and offset in calendar